어드레스레스트풀 딜리트쪽 데이터타입변경(long->int)

+++ backend/src/main/java/spring/boot/backend/service/AddressRestService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface AddressRestService {
 	List<Address> listS();
 	void insertS(Address address);
-	void deleteS(long seq);
+	void deleteS(int seq);
 	//for Ajax
 	Address selectBySeqS(long seq);
 	List<Address> selectByNameS(String name);
